Reapply "merged 3D and 2D, changed to functional components"
This reverts commit 53053d5baaf3e54c250272fb9ed62974da11d769.
