--- frontend/src/2D/styles.scss
@@ -1,14 +1,0 @@
-.TwoD-wrapper {
-    padding: 20px;
-    background-color: #182B49;
-    height: 100vh;
-    color: #fff;
-
-    .intro-card {
-        background-color: #fff;
-        border-radius: 3px;
-        color: #000;
-        padding: 10px;
-    }
-}
-
