--- frontend/src/3D/styles.scss
@@ -1,15 +1,0 @@
-.ThreeD-wrapper {
-    padding: 20px;
-    background-color: #182B49;
-    height: 100vh;
-    width: 100vw;
-    color: #fff;
-
-    .intro-card {
-        background-color: #fff;
-        width: 600px;
-        border-radius: 3px;
-        color: #000;
-        padding: 10px;
-    }
-}
