+++ frontend/src/App.scss
@@ -38,3 +38,33 @@
     transform: rotate(360deg);
   }
 }
+
+.TwoD-wrapper {
+  padding: 20px;
+  background-color: #182B49;
+  height: 100vh;
+  color: #fff;
+
+  .intro-card {
+    background-color: #fff;
+    border-radius: 3px;
+    color: #000;
+    padding: 10px;
+  }
+}
+
+.ThreeD-wrapper {
+  // padding: 20px;
+  // width: '80%';
+  // height: 'auto';
+  // background-color: #182B49;
+  // color: #fff;
+
+  // .intro-card {
+  //   background-color: #fff;
+  //   width: 600px;
+  //   border-radius: 3px;
+  //   color: #000;
+  //   padding: 10px;
+  // }
+}
