+++ frontend/src/models/STLModel.js
